+++ TM/ptrh.h
@@ -5,7 +5,7 @@
 #include "nl_assert.h"
 #include "subbus.h"
 
-typedef unsigned short USHRT;
+typedef uint16_t USHRT;
 class ptrh {
   public:
     ptrh( const char *id, USHRT base, USHRT C1, USHRT C2,
@@ -27,7 +27,7 @@ class ptrh {
     USHRT base_addr;
     USHRT C[6];
     double C1d, C2d, C3d, C4d, C6d;
-    unsigned long C5d;
+    uint32_t C5d;
     double Off, Sens;
     int check_coeff( int i, USHRT C);
     int stale;
